Spring2024/B657/Assignment1/pyproject.toml
[tool.poetry]
name = "bhnaraya-nhaldert-hkushwah-a1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.2.0"
numpy = "^1.26.4"
tqdm = "^4.66.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
